HCSVLAB-1014: comment out mcr installer dependency
The current version of the MCR installer does not match the version used to compile pysound3, so a prior (correctly versioned) installation is assumed.
Once the correct version of the MCR Installer is located, the commented code can be used to ensure that the toolshed can add and remove the MCR as part of the tool installation and uninstallation.