MISC: update galaxy tools with manual install requirements
Modified jc parser and psysyound readme to include all of the required manual installation steps as instructions. These README files are automatically displayed in Galaxy when selecting to install a tool from the toolshed. Also included a setup script for psysound which mirrors the steps included in the README file.